Replace m4 with python script parsing TSV
While POSIX compliant, m4 is a massive fucking headache. The quoting.
The arguments. The comments. The quoting. The flow. The quoting. It's
just not worth it.

For more complex templating needs, using a SSG like Hugo or Kiln would
be a better choice. But for my use this would be a bit much, as I only
have one file and relatively simple styling.

Speaking of styling, CSS classes were expanded upon to make it easier to
give different tiers of entries specific styling. This was a problem
with the previous system of nested m4 macros because it didn't allow for
more specialized styling within each entry: for instance, changing the
color of the link specifically of an entry.

This commit also adds an entry for Everything Everywhere All at Once.